requirements.txt
newscatcherapi
Python-dotenv
Flask
streamlit
accelerate
torch
transformers
huggingface-hub
langchain
bs4
rank_bm25
sentencepiece